+++ example/gzip-transform/gunzip.c
@@ -167,7 +167,7 @@ gzip_transform_one(GzipData * data, INKIOBufferReader input_reader, int amount)
   INKIOBufferBlock blkp;
   const char *ibuf;
   char *obuf;
-  int ilength, olength;
+  int64 ilength, olength;
   int err = Z_OK;
 
   while (amount > 0) {
@@ -236,7 +236,7 @@ gzip_transform_finish(GzipData * data)
   if (data->state == 1) {
     INKIOBufferBlock blkp;
     char *obuf;
-    int olength;
+    int64 olength;
     int err;
 
     data->state = 2;
